+++ app/overrides/custom_fields/index.rb
@@ -2,7 +2,7 @@
 #
 # Redmine plugin for xmera called Project Types Plugin.
 #
-# Copyright (C) 2017-21 Liane Hampe <liaham@xmera.de>, xmera.
+# Copyright (C) 2021 Liane Hampe <liaham@xmera.de>, xmera.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -18,22 +18,26 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-# Target is redmines app/views/trackers/_form.html.erb file
+# Target is redmines app/views/custom_fields/_form.html.erb file
+# Removes the list of projects from the custom_fields form since the the assignment
+# to a specific project is already defined by the assingment to a tracker which is 
+# still possible.
+
 Deface::Override.new(
-  virtual_path: 'trackers/_form',
-  name: 'remove-project-list',
-  replace: "div.splitcontentright",
-  partial: "trackers/project_types",
+  virtual_path: 'custom_fields/_index',
+  name: 'disables-project-specific-table-columns',
+  remove: "erb[silent]:contains('IssueCustomField')",
+  closing_selector: "erb[silent]:contains('end')",
+  original: 'f7d6b30d173528f6f51a99565b04b9560ab75706',
   disabled: ProjectTypes.missing?,
-  original: '570b247e7ac59c8c751282f89dd9011b7fe6f945',
   namespaced: true
 )
 Deface::Override.new(
-  virtual_path: 'trackers/_form',
-  name: 'remove-custom-field-list',
-  remove: "erb[silent]:contains('if @issue_custom_fields.present?')",
+  virtual_path: 'custom_fields/_index',
+  name: 'disables-project-specific-table-entries',
+  remove: "erb[silent]:contains('if tab[:name]')",
   closing_selector: "erb[silent]:contains('end')",
+  original: '3f3eee8b0b83eda2fade16a1a5a70027476470c7',
   disabled: ProjectTypes.missing?,
-  original: 'f78ee40e91e9a5c255510544e02182d9c3da046d',
   namespaced: true
 )