+++ config/routes.rb
@@ -21,4 +21,4 @@
 # Plugin's routes
 # See: http://guides.rubyonrails.org/routing.html
 
- resources :project_types#, :project_types_default_modules, :projects_project_types
+ resources :project_types