+++ lib/project_types/patches/custom_field_patch.rb
@@ -21,50 +21,28 @@
 module ProjectTypes
   module Patches
     module CustomFieldPatch
-      # Calls the prepended modules on the class. Without this 
-      # method the after_commit would be called on the module
-      # what raises an error since the method is unknown for the
-      # module IssueCloning::Patches::RolePatch.
-      def self.prepended(mod)
-         mod.singleton_class.prepend(ClassMethods)
-         mod.prepend(InstanceMethods)
-         mod.after_commit :sync_project_custom_fields
+      def self.prepended(base)
+        base.extend(ClassMethods)
+        base.include(InstanceMethods) 
+        base.class_eval do
+          include ProjectTypes::Switch::IssueCustomFields
+                
+          after_initialize do |issue_custom_field|
+            enable_switch(:issue_custom_fields) if ProjectTypes.any?
+          end
+        end
       end      
       
       module ClassMethods
-        # empty
+        def enable_switch(name)
+          send name
+        end
       end
 
-      module InstanceMethods 
-        private
-        # The assignment of custom fields to projects by the user is not displayed
-        # anymore. See app/overrides/custom_fields/form.
-        # Instead the assignment is executed automatically in background based
-        # on the tracker.
-        # The relation is project -> project type -> tracker <-custom field.
-        # That is, with every changing in the relation between custom fields and tracker
-        # the relation between custom fields and projects needs to be synchronised.
-        # The whole chain is: project -> project type -> tracker <-> custom field.
-        # The following method defines: project -> project type -> tracker <- custom field.
-        # 
-        def sync_project_custom_fields
-          # There are several types of custom fields. Relevant are only IssueCustomFields.
-          # Each project has many trackers and many custom fields.
-          # The relation project -> tracker (p.tracker_ids) is maintained by project_types_default_tracker.rb.
-          # Therefore, p.tracker_ids is reliable.
-          if ["IssueCustomField"].include?(self.class.to_s) && ProjectType.any?     
-            Project.all.each do |p|
-              intersection = p.tracker_ids & self.tracker_ids
-              if intersection.empty?
-                # Check whether to delete previously assigned custom fields of the project
-                self.project_ids.delete(p.id)
-              else
-                # Update of projects custom fields
-                self.project_ids += [p.id] unless self.project_ids.include?(p.id)
-              end
-            end if Project.any?
-          end
-        end
+      module InstanceMethods
+        def enable_switch(name)
+          self.class.enable_switch(name)
+        end  
       end
     end
   end