+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2021-02-07
+
+### Changed
+
+* all ProjectType model associations and its tables
+* ProjectType views to be more consistend with Redmine
+
+### Added
+
+* further tests
+
 ## 1.0.1 - 2021-01-26
 
 ### Added